app/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["vjanz <valon.januzaj98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.60.1"
requests = "^2.24.0"
uvicorn = "^0.11.7"
alembic = "^1.4.2"
celery = "^4.4.6"
redis = "^3.5.3"
httpx = "^0.13.3"
pytest = "^6.0.1"
SQLAlchemy = "^1.3.18"
starlette = "^0.13.6"
passlib = "^1.7.2"
bcrypt = "^3.1.7"
sqlalchemy-utils = "^0.36.8"
python-multipart = "^0.0.5"
pyjwt = "^1.7.1"
psycopg2 = "^2.8.5"
databases = "^0.3.2"
starlette_exporter = "^0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
